chore: merge test_states into test_data
